beatmap info refactor pt1
