+++ src/mixins.js
@@ -1,24 +1,24 @@
 const mixin = {
     methods: {
-        executePostT: async function(path, data, e) {
-			if (e) e.target.disabled = true;
+		async executePost (path, data, e) {
+            if (e) e.target.disabled = true;
+            $("[data-toggle='tooltip']").tooltip('hide');
 
-			try {
-				const res = await axios.post(path, data);
-				
-				if (res.data.error) {
+            try {
+                const res = await axios.post(path, data);
+                if (res.data.error) {
                     this.info = res.data.error;
-				} else {
-					if (e) e.target.disabled = false;
-					return res.data;
-				}
-			} catch (error) {
-				console.log(error)
-			}
-			
-			if (e) e.target.disabled = false;
-		},
-		//sort
+                    this.inviteConfirm = null;
+                } else {
+                    if (e) e.target.disabled = false;
+                    return res.data;
+                }
+            } catch (error) {
+                this.info = 'Something went wrong';
+            }
+            
+            if (e) e.target.disabled = false;
+        },
     }
 }
